YE-3091 (FIX) Getting rid of proxy via OMM
When disabling the kartina channels for a room in OAM, its not reflexed in the player.

Further more: The (Wolfgang Haupt, Lukas Obermann) the player keeps rebooting. The player prob. receives a faulty input, but it should never crash because of that. Eventually it shows some red coloured Kodi page on the player.

The system works now like this: (which is horrible!) In OAM you disable Kartina for a room. This will trigger the player for that room to fetch is new config. This request will go to OMM. OMM will 'proxy-pass' this request to OAM with the calling ip of the player. OAM will filter the result based on the room config (in this case it will filter-out the Kartina channels).